% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglayer.R
\name{gglayer}
\alias{gglayer}
\title{Add a geom to all plots}
\usage{
gglayer(.data, layer_use, ...)
}
\arguments{
\item{.data}{Data frame with nested data column}

\item{layer_use}{The ggplot2 geom that you want to add}

\item{...}{Other arguments passed onto the geom specified in \code{layer_use}}
}
\description{
This function adds a geom to the plots contained within the column \code{plot}.
}
\examples{
eg <- mtcars \%>\%
  group_by(am) \%>\%
  nest('data'=-am)\%>\%
  mutate('colour'=ifelse(am==0,'red','purple'),
         'title'=sprintf('AM: \%i',am),
         'subtitle'=sprintf('Data from \%i different models',map_dbl(data,nrow)),
         'minlimit'=ifelse(am==0,-1000,-250),
         'maxlimit'=map_dbl(data,function(x)max(x$disp)))

out <- eg \%>\%
  ggpipe("data") \%>\%
  gglayer(geom_point,aes(x=disp,y=hp),colour=colour)\%>\%
  gglayer(scale_x_continuous,limits=c(minlimit,maxlimit))\%>\%
  gglayer(labs,title=title)\%>\%
  gglayer(facet_wrap,cyl~.)\%>\%
  ggtheme(theme_bw())

out$plot
}
