% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa.R
\docType{package}
\name{ggpipe-package}
\alias{ggpipe}
\alias{ggpipe-package}
\title{ggplot with the pipe}
\description{
Use ggplot functions with the pipe instead of the "+" operator
}
\examples{
ggplot(mtcars) \%>\% geom_point(aes(mpg, cyl))
}
\author{
\strong{Maintainer}: Sergio Oller \email{sergioller@gmail.com}

}
