% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa.R
\docType{data}
\name{ggplot-pipeified}
\alias{ggplot-pipeified}
\alias{.ggplot_pipeified_funs}
\alias{aes}
\alias{aes_}
\alias{aes_all}
\alias{aes_auto}
\alias{aes_q}
\alias{aes_string}
\alias{annotate}
\alias{coord_cartesian}
\alias{coord_equal}
\alias{coord_fixed}
\alias{coord_flip}
\alias{coord_map}
\alias{coord_munch}
\alias{coord_polar}
\alias{coord_quickmap}
\alias{coord_trans}
\alias{facet_grid}
\alias{facet_null}
\alias{facet_wrap}
\alias{geom_abline}
\alias{geom_area}
\alias{geom_bar}
\alias{geom_bin2d}
\alias{geom_blank}
\alias{geom_boxplot}
\alias{geom_col}
\alias{geom_contour}
\alias{geom_count}
\alias{geom_crossbar}
\alias{geom_curve}
\alias{geom_density}
\alias{geom_density2d}
\alias{geom_density_2d}
\alias{geom_dotplot}
\alias{geom_errorbar}
\alias{geom_errorbarh}
\alias{geom_freqpoly}
\alias{geom_hex}
\alias{geom_histogram}
\alias{geom_hline}
\alias{geom_jitter}
\alias{geom_label}
\alias{geom_line}
\alias{geom_linerange}
\alias{geom_map}
\alias{geom_path}
\alias{geom_point}
\alias{geom_pointrange}
\alias{geom_polygon}
\alias{geom_qq}
\alias{geom_quantile}
\alias{geom_raster}
\alias{geom_rect}
\alias{geom_ribbon}
\alias{geom_rug}
\alias{geom_segment}
\alias{geom_smooth}
\alias{geom_spoke}
\alias{geom_step}
\alias{geom_text}
\alias{geom_tile}
\alias{geom_violin}
\alias{geom_vline}
\alias{ggtitle}
\alias{labs}
\alias{scale_alpha}
\alias{scale_alpha_continuous}
\alias{scale_alpha_discrete}
\alias{scale_alpha_identity}
\alias{scale_alpha_manual}
\alias{scale_color_brewer}
\alias{scale_color_continuous}
\alias{scale_color_discrete}
\alias{scale_color_distiller}
\alias{scale_color_gradient}
\alias{scale_color_gradient2}
\alias{scale_color_gradientn}
\alias{scale_color_grey}
\alias{scale_color_hue}
\alias{scale_color_identity}
\alias{scale_color_manual}
\alias{scale_colour_brewer}
\alias{scale_colour_continuous}
\alias{scale_colour_date}
\alias{scale_colour_datetime}
\alias{scale_colour_discrete}
\alias{scale_colour_distiller}
\alias{scale_colour_gradient}
\alias{scale_colour_gradient2}
\alias{scale_colour_gradientn}
\alias{scale_colour_grey}
\alias{scale_colour_hue}
\alias{scale_colour_identity}
\alias{scale_colour_manual}
\alias{scale_fill_brewer}
\alias{scale_fill_continuous}
\alias{scale_fill_date}
\alias{scale_fill_datetime}
\alias{scale_fill_discrete}
\alias{scale_fill_distiller}
\alias{scale_fill_gradient}
\alias{scale_fill_gradient2}
\alias{scale_fill_gradientn}
\alias{scale_fill_grey}
\alias{scale_fill_hue}
\alias{scale_fill_identity}
\alias{scale_fill_manual}
\alias{scale_linetype}
\alias{scale_linetype_continuous}
\alias{scale_linetype_discrete}
\alias{scale_linetype_identity}
\alias{scale_linetype_manual}
\alias{scale_radius}
\alias{scale_shape}
\alias{scale_shape_continuous}
\alias{scale_shape_discrete}
\alias{scale_shape_identity}
\alias{scale_shape_manual}
\alias{scale_size}
\alias{scale_size_area}
\alias{scale_size_continuous}
\alias{scale_size_date}
\alias{scale_size_datetime}
\alias{scale_size_discrete}
\alias{scale_size_identity}
\alias{scale_size_manual}
\alias{scale_x_continuous}
\alias{scale_x_date}
\alias{scale_x_datetime}
\alias{scale_x_discrete}
\alias{scale_x_log10}
\alias{scale_x_reverse}
\alias{scale_x_sqrt}
\alias{scale_x_time}
\alias{scale_y_continuous}
\alias{scale_y_date}
\alias{scale_y_datetime}
\alias{scale_y_discrete}
\alias{scale_y_log10}
\alias{scale_y_reverse}
\alias{scale_y_sqrt}
\alias{scale_y_time}
\alias{stat_bin}
\alias{stat_bin2d}
\alias{stat_bin_2d}
\alias{stat_binhex}
\alias{stat_bin_hex}
\alias{stat_boxplot}
\alias{stat_contour}
\alias{stat_count}
\alias{stat_density}
\alias{stat_density2d}
\alias{stat_density_2d}
\alias{stat_ecdf}
\alias{stat_ellipse}
\alias{stat_function}
\alias{stat_identity}
\alias{stat_qq}
\alias{stat_quantile}
\alias{stat_smooth}
\alias{stat_spoke}
\alias{stat_sum}
\alias{stat_summary}
\alias{stat_summary2d}
\alias{stat_summary_2d}
\alias{stat_summary_bin}
\alias{stat_summary_hex}
\alias{stat_unique}
\alias{stat_ydensity}
\alias{theme}
\alias{theme_bw}
\alias{theme_classic}
\alias{theme_dark}
\alias{theme_get}
\alias{theme_gray}
\alias{theme_grey}
\alias{theme_light}
\alias{theme_linedraw}
\alias{theme_minimal}
\alias{theme_replace}
\alias{theme_set}
\alias{theme_update}
\alias{theme_void}
\alias{xlab}
\alias{ylab}
\title{ggplot2 functions modified to work with the pipe}
\format{An object of class \code{NULL} of length 0.}
\usage{
.ggplot_pipeified_funs
}
\description{
ggplot2 functions modified to work with the pipe
}
\details{
These functions are imported from ggplot2. They are adapted to
have a first argument called "gplt", that is the \code{ggplot}
object. Follow the links below to see their documentation.
\describe{
  \item{aes}{\code{\link[ggplot2]{aes}}}
  \item{aes_}{\code{\link[ggplot2]{aes_}}}
  \item{aes_all}{\code{\link[ggplot2]{aes_all}}}
  \item{aes_auto}{\code{\link[ggplot2]{aes_auto}}}
  \item{aes_q}{\code{\link[ggplot2]{aes_q}}}
  \item{aes_string}{\code{\link[ggplot2]{aes_string}}}
  \item{annotate}{\code{\link[ggplot2]{annotate}}}
  \item{coord_cartesian}{\code{\link[ggplot2]{coord_cartesian}}}
  \item{coord_equal}{\code{\link[ggplot2]{coord_equal}}}
  \item{coord_fixed}{\code{\link[ggplot2]{coord_fixed}}}
  \item{coord_flip}{\code{\link[ggplot2]{coord_flip}}}
  \item{coord_map}{\code{\link[ggplot2]{coord_map}}}
  \item{coord_munch}{\code{\link[ggplot2]{coord_munch}}}
  \item{coord_polar}{\code{\link[ggplot2]{coord_polar}}}
  \item{coord_quickmap}{\code{\link[ggplot2]{coord_quickmap}}}
  \item{coord_trans}{\code{\link[ggplot2]{coord_trans}}}
  \item{facet_grid}{\code{\link[ggplot2]{facet_grid}}}
  \item{facet_null}{\code{\link[ggplot2]{facet_null}}}
  \item{facet_wrap}{\code{\link[ggplot2]{facet_wrap}}}
  \item{geom_abline}{\code{\link[ggplot2]{geom_abline}}}
  \item{geom_area}{\code{\link[ggplot2]{geom_area}}}
  \item{geom_bar}{\code{\link[ggplot2]{geom_bar}}}
  \item{geom_bin2d}{\code{\link[ggplot2]{geom_bin2d}}}
  \item{geom_blank}{\code{\link[ggplot2]{geom_blank}}}
  \item{geom_boxplot}{\code{\link[ggplot2]{geom_boxplot}}}
  \item{geom_col}{\code{\link[ggplot2]{geom_col}}}
  \item{geom_contour}{\code{\link[ggplot2]{geom_contour}}}
  \item{geom_count}{\code{\link[ggplot2]{geom_count}}}
  \item{geom_crossbar}{\code{\link[ggplot2]{geom_crossbar}}}
  \item{geom_curve}{\code{\link[ggplot2]{geom_curve}}}
  \item{geom_density}{\code{\link[ggplot2]{geom_density}}}
  \item{geom_density2d}{\code{\link[ggplot2]{geom_density2d}}}
  \item{geom_density_2d}{\code{\link[ggplot2]{geom_density_2d}}}
  \item{geom_dotplot}{\code{\link[ggplot2]{geom_dotplot}}}
  \item{geom_errorbar}{\code{\link[ggplot2]{geom_errorbar}}}
  \item{geom_errorbarh}{\code{\link[ggplot2]{geom_errorbarh}}}
  \item{geom_freqpoly}{\code{\link[ggplot2]{geom_freqpoly}}}
  \item{geom_hex}{\code{\link[ggplot2]{geom_hex}}}
  \item{geom_histogram}{\code{\link[ggplot2]{geom_histogram}}}
  \item{geom_hline}{\code{\link[ggplot2]{geom_hline}}}
  \item{geom_jitter}{\code{\link[ggplot2]{geom_jitter}}}
  \item{geom_label}{\code{\link[ggplot2]{geom_label}}}
  \item{geom_line}{\code{\link[ggplot2]{geom_line}}}
  \item{geom_linerange}{\code{\link[ggplot2]{geom_linerange}}}
  \item{geom_map}{\code{\link[ggplot2]{geom_map}}}
  \item{geom_path}{\code{\link[ggplot2]{geom_path}}}
  \item{geom_point}{\code{\link[ggplot2]{geom_point}}}
  \item{geom_pointrange}{\code{\link[ggplot2]{geom_pointrange}}}
  \item{geom_polygon}{\code{\link[ggplot2]{geom_polygon}}}
  \item{geom_qq}{\code{\link[ggplot2]{geom_qq}}}
  \item{geom_quantile}{\code{\link[ggplot2]{geom_quantile}}}
  \item{geom_raster}{\code{\link[ggplot2]{geom_raster}}}
  \item{geom_rect}{\code{\link[ggplot2]{geom_rect}}}
  \item{geom_ribbon}{\code{\link[ggplot2]{geom_ribbon}}}
  \item{geom_rug}{\code{\link[ggplot2]{geom_rug}}}
  \item{geom_segment}{\code{\link[ggplot2]{geom_segment}}}
  \item{geom_smooth}{\code{\link[ggplot2]{geom_smooth}}}
  \item{geom_spoke}{\code{\link[ggplot2]{geom_spoke}}}
  \item{geom_step}{\code{\link[ggplot2]{geom_step}}}
  \item{geom_text}{\code{\link[ggplot2]{geom_text}}}
  \item{geom_tile}{\code{\link[ggplot2]{geom_tile}}}
  \item{geom_violin}{\code{\link[ggplot2]{geom_violin}}}
  \item{geom_vline}{\code{\link[ggplot2]{geom_vline}}}
  \item{ggtitle}{\code{\link[ggplot2]{ggtitle}}}
  \item{labs}{\code{\link[ggplot2]{labs}}}
  \item{scale_alpha}{\code{\link[ggplot2]{scale_alpha}}}
  \item{scale_alpha_continuous}{\code{\link[ggplot2]{scale_alpha_continuous}}}
  \item{scale_alpha_discrete}{\code{\link[ggplot2]{scale_alpha_discrete}}}
  \item{scale_alpha_identity}{\code{\link[ggplot2]{scale_alpha_identity}}}
  \item{scale_alpha_manual}{\code{\link[ggplot2]{scale_alpha_manual}}}
  \item{scale_color_brewer}{\code{\link[ggplot2]{scale_color_brewer}}}
  \item{scale_color_continuous}{\code{\link[ggplot2]{scale_color_continuous}}}
  \item{scale_color_discrete}{\code{\link[ggplot2]{scale_color_discrete}}}
  \item{scale_color_distiller}{\code{\link[ggplot2]{scale_color_distiller}}}
  \item{scale_color_gradient}{\code{\link[ggplot2]{scale_color_gradient}}}
  \item{scale_color_gradient2}{\code{\link[ggplot2]{scale_color_gradient2}}}
  \item{scale_color_gradientn}{\code{\link[ggplot2]{scale_color_gradientn}}}
  \item{scale_color_grey}{\code{\link[ggplot2]{scale_color_grey}}}
  \item{scale_color_hue}{\code{\link[ggplot2]{scale_color_hue}}}
  \item{scale_color_identity}{\code{\link[ggplot2]{scale_color_identity}}}
  \item{scale_color_manual}{\code{\link[ggplot2]{scale_color_manual}}}
  \item{scale_colour_brewer}{\code{\link[ggplot2]{scale_colour_brewer}}}
  \item{scale_colour_continuous}{\code{\link[ggplot2]{scale_colour_continuous}}}
  \item{scale_colour_date}{\code{\link[ggplot2]{scale_colour_date}}}
  \item{scale_colour_datetime}{\code{\link[ggplot2]{scale_colour_datetime}}}
  \item{scale_colour_discrete}{\code{\link[ggplot2]{scale_colour_discrete}}}
  \item{scale_colour_distiller}{\code{\link[ggplot2]{scale_colour_distiller}}}
  \item{scale_colour_gradient}{\code{\link[ggplot2]{scale_colour_gradient}}}
  \item{scale_colour_gradient2}{\code{\link[ggplot2]{scale_colour_gradient2}}}
  \item{scale_colour_gradientn}{\code{\link[ggplot2]{scale_colour_gradientn}}}
  \item{scale_colour_grey}{\code{\link[ggplot2]{scale_colour_grey}}}
  \item{scale_colour_hue}{\code{\link[ggplot2]{scale_colour_hue}}}
  \item{scale_colour_identity}{\code{\link[ggplot2]{scale_colour_identity}}}
  \item{scale_colour_manual}{\code{\link[ggplot2]{scale_colour_manual}}}
  \item{scale_fill_brewer}{\code{\link[ggplot2]{scale_fill_brewer}}}
  \item{scale_fill_continuous}{\code{\link[ggplot2]{scale_fill_continuous}}}
  \item{scale_fill_date}{\code{\link[ggplot2]{scale_fill_date}}}
  \item{scale_fill_datetime}{\code{\link[ggplot2]{scale_fill_datetime}}}
  \item{scale_fill_discrete}{\code{\link[ggplot2]{scale_fill_discrete}}}
  \item{scale_fill_distiller}{\code{\link[ggplot2]{scale_fill_distiller}}}
  \item{scale_fill_gradient}{\code{\link[ggplot2]{scale_fill_gradient}}}
  \item{scale_fill_gradient2}{\code{\link[ggplot2]{scale_fill_gradient2}}}
  \item{scale_fill_gradientn}{\code{\link[ggplot2]{scale_fill_gradientn}}}
  \item{scale_fill_grey}{\code{\link[ggplot2]{scale_fill_grey}}}
  \item{scale_fill_hue}{\code{\link[ggplot2]{scale_fill_hue}}}
  \item{scale_fill_identity}{\code{\link[ggplot2]{scale_fill_identity}}}
  \item{scale_fill_manual}{\code{\link[ggplot2]{scale_fill_manual}}}
  \item{scale_linetype}{\code{\link[ggplot2]{scale_linetype}}}
  \item{scale_linetype_continuous}{\code{\link[ggplot2]{scale_linetype_continuous}}}
  \item{scale_linetype_discrete}{\code{\link[ggplot2]{scale_linetype_discrete}}}
  \item{scale_linetype_identity}{\code{\link[ggplot2]{scale_linetype_identity}}}
  \item{scale_linetype_manual}{\code{\link[ggplot2]{scale_linetype_manual}}}
  \item{scale_radius}{\code{\link[ggplot2]{scale_radius}}}
  \item{scale_shape}{\code{\link[ggplot2]{scale_shape}}}
  \item{scale_shape_continuous}{\code{\link[ggplot2]{scale_shape_continuous}}}
  \item{scale_shape_discrete}{\code{\link[ggplot2]{scale_shape_discrete}}}
  \item{scale_shape_identity}{\code{\link[ggplot2]{scale_shape_identity}}}
  \item{scale_shape_manual}{\code{\link[ggplot2]{scale_shape_manual}}}
  \item{scale_size}{\code{\link[ggplot2]{scale_size}}}
  \item{scale_size_area}{\code{\link[ggplot2]{scale_size_area}}}
  \item{scale_size_continuous}{\code{\link[ggplot2]{scale_size_continuous}}}
  \item{scale_size_date}{\code{\link[ggplot2]{scale_size_date}}}
  \item{scale_size_datetime}{\code{\link[ggplot2]{scale_size_datetime}}}
  \item{scale_size_discrete}{\code{\link[ggplot2]{scale_size_discrete}}}
  \item{scale_size_identity}{\code{\link[ggplot2]{scale_size_identity}}}
  \item{scale_size_manual}{\code{\link[ggplot2]{scale_size_manual}}}
  \item{scale_x_continuous}{\code{\link[ggplot2]{scale_x_continuous}}}
  \item{scale_x_date}{\code{\link[ggplot2]{scale_x_date}}}
  \item{scale_x_datetime}{\code{\link[ggplot2]{scale_x_datetime}}}
  \item{scale_x_discrete}{\code{\link[ggplot2]{scale_x_discrete}}}
  \item{scale_x_log10}{\code{\link[ggplot2]{scale_x_log10}}}
  \item{scale_x_reverse}{\code{\link[ggplot2]{scale_x_reverse}}}
  \item{scale_x_sqrt}{\code{\link[ggplot2]{scale_x_sqrt}}}
  \item{scale_x_time}{\code{\link[ggplot2]{scale_x_time}}}
  \item{scale_y_continuous}{\code{\link[ggplot2]{scale_y_continuous}}}
  \item{scale_y_date}{\code{\link[ggplot2]{scale_y_date}}}
  \item{scale_y_datetime}{\code{\link[ggplot2]{scale_y_datetime}}}
  \item{scale_y_discrete}{\code{\link[ggplot2]{scale_y_discrete}}}
  \item{scale_y_log10}{\code{\link[ggplot2]{scale_y_log10}}}
  \item{scale_y_reverse}{\code{\link[ggplot2]{scale_y_reverse}}}
  \item{scale_y_sqrt}{\code{\link[ggplot2]{scale_y_sqrt}}}
  \item{scale_y_time}{\code{\link[ggplot2]{scale_y_time}}}
  \item{stat_bin}{\code{\link[ggplot2]{stat_bin}}}
  \item{stat_bin2d}{\code{\link[ggplot2]{stat_bin2d}}}
  \item{stat_bin_2d}{\code{\link[ggplot2]{stat_bin_2d}}}
  \item{stat_binhex}{\code{\link[ggplot2]{stat_binhex}}}
  \item{stat_bin_hex}{\code{\link[ggplot2]{stat_bin_hex}}}
  \item{stat_boxplot}{\code{\link[ggplot2]{stat_boxplot}}}
  \item{stat_contour}{\code{\link[ggplot2]{stat_contour}}}
  \item{stat_count}{\code{\link[ggplot2]{stat_count}}}
  \item{stat_density}{\code{\link[ggplot2]{stat_density}}}
  \item{stat_density2d}{\code{\link[ggplot2]{stat_density2d}}}
  \item{stat_density_2d}{\code{\link[ggplot2]{stat_density_2d}}}
  \item{stat_ecdf}{\code{\link[ggplot2]{stat_ecdf}}}
  \item{stat_ellipse}{\code{\link[ggplot2]{stat_ellipse}}}
  \item{stat_function}{\code{\link[ggplot2]{stat_function}}}
  \item{stat_identity}{\code{\link[ggplot2]{stat_identity}}}
  \item{stat_qq}{\code{\link[ggplot2]{stat_qq}}}
  \item{stat_quantile}{\code{\link[ggplot2]{stat_quantile}}}
  \item{stat_smooth}{\code{\link[ggplot2]{stat_smooth}}}
  \item{stat_spoke}{\code{\link[ggplot2]{stat_spoke}}}
  \item{stat_sum}{\code{\link[ggplot2]{stat_sum}}}
  \item{stat_summary}{\code{\link[ggplot2]{stat_summary}}}
  \item{stat_summary2d}{\code{\link[ggplot2]{stat_summary2d}}}
  \item{stat_summary_2d}{\code{\link[ggplot2]{stat_summary_2d}}}
  \item{stat_summary_bin}{\code{\link[ggplot2]{stat_summary_bin}}}
  \item{stat_summary_hex}{\code{\link[ggplot2]{stat_summary_hex}}}
  \item{stat_unique}{\code{\link[ggplot2]{stat_unique}}}
  \item{stat_ydensity}{\code{\link[ggplot2]{stat_ydensity}}}
  \item{theme}{\code{\link[ggplot2]{theme}}}
  \item{theme_bw}{\code{\link[ggplot2]{theme_bw}}}
  \item{theme_classic}{\code{\link[ggplot2]{theme_classic}}}
  \item{theme_dark}{\code{\link[ggplot2]{theme_dark}}}
  \item{theme_get}{\code{\link[ggplot2]{theme_get}}}
  \item{theme_gray}{\code{\link[ggplot2]{theme_gray}}}
  \item{theme_grey}{\code{\link[ggplot2]{theme_grey}}}
  \item{theme_light}{\code{\link[ggplot2]{theme_light}}}
  \item{theme_linedraw}{\code{\link[ggplot2]{theme_linedraw}}}
  \item{theme_minimal}{\code{\link[ggplot2]{theme_minimal}}}
  \item{theme_replace}{\code{\link[ggplot2]{theme_replace}}}
  \item{theme_set}{\code{\link[ggplot2]{theme_set}}}
  \item{theme_update}{\code{\link[ggplot2]{theme_update}}}
  \item{theme_void}{\code{\link[ggplot2]{theme_void}}}
  \item{xlab}{\code{\link[ggplot2]{xlab}}}
  \item{ylab}{\code{\link[ggplot2]{ylab}}}
}
}
\keyword{internal}
