% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa.R
\name{ggsave}
\alias{ggsave}
\title{Save a ggplot}
\usage{
ggsave(filename, plot, ...)
}
\arguments{
\item{...}{Arguments passed on to \code{ggplot2::ggsave}
\describe{
  \item{filename}{File name to create on disk.}
  \item{plot}{Plot to save, defaults to last plot displayed.}
  \item{device}{Device to use. Can be either be a device function
(e.g. \code{\link{png}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}
  \item{path}{Path to save plot to (combined with filename).}
  \item{scale}{Multiplicative scaling factor.}
  \item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
  \item{height}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
  \item{units}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
  \item{dpi}{Plot resolution. Applies only to raster output types.}
  \item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
}}
}
\description{
With respect to [ggplot2::ggsave()], this function checks the types
and swaps the filename and plot arguments if needed, so it is pipeable
}
\seealso{
ggplot2::ggsave
}
