% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{ggtheme}
\alias{ggtheme}
\title{Add a theme to all plots}
\usage{
ggtheme(.data, theme)
}
\arguments{
\item{.data}{Data frame with nested data column}

\item{theme}{The ggplot2 theme you want to add}
}
\description{
This function adds a theme to the plots contained within the column \code{plot}
}
