% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa.R
\name{unggplot}
\alias{unggplot}
\title{Returns the data from the ggplot}
\usage{
unggplot(gplt, layer = NULL)
}
\arguments{
\item{gplt}{A ggplot object}

\item{layer}{ggplot plots can include different data in other layers. This is
the index of the layer to return (by default the global data
is returned)}
}
\value{
a data frame
}
\description{
Returns the data frame from the \code{gplt} object
}
\examples{

data(iris)
ggplot(iris) \%>\%
  geom_point(aes(x = Sepal.Length, y = Sepal.Width, color = Species)) \%>\%
  ggsave("demo.png") \%>\%
  unggplot() \%>\%
  head()
}
